m/requirements.txt
torch
transformers
sentencepiece
streamlit
fastapi
uvicorn
pydantic
langdetect
babel
pyyaml
tqdm